WIP move to Graph structure that abstracts away PKM differences
